integrated jacoco plug-in and added required JUnit tests for 100% Test Coverage for controller, model and services

+++ src/main/java/com/iheart/ms/repository/AdvertiserRepository.java
@@ -21,10 +21,10 @@ public interface AdvertiserRepository {
 		public Collection<Advertiser> findAll();
 
 		@Select("SELECT * FROM advertiser WHERE id = #{id}")
-		public Advertiser findById(int id);
+		public Advertiser findById(Long id);
 
 		@Delete("DELETE FROM advertiser WHERE id = #{id}")
-		public int deleteById(int id);
+		public int deleteById(Long id);
 
 		@Options(useGeneratedKeys=true,keyProperty="id",  keyColumn="id")
 		@Insert("INSERT INTO advertiser(contact_name, credit_limit) VALUES (#{contactName}, #{creditLimit})")
